chore: address pr feedback
