+++ src/lib/schematics/address-form/schema.json
@@ -1,7 +1,7 @@
 {
   "$schema": "http://json-schema.org/schema",
-  "id": "SchematicsMaterialForm",
-  "title": "Material Form Options Schema",
+  "id": "SchematicsMaterialAddressForm",
+  "title": "Material Address Form Options Schema",
   "type": "object",
   "properties": {
     "path": {
