+++ src/lib/schematics/collection.json
@@ -32,11 +32,11 @@
       "aliases": [ "material-nav" ]
     },
     // Creates a address form component
-    "form": {
+    "addressForm": {
       "description": "Create a component with a address form",
-      "factory": "./form/index",
-      "schema": "./form/schema.json",
-      "aliases": []
+      "factory": "./address-form/index",
+      "schema": "./address-form/schema.json",
+      "aliases": ["address-form"]
     }
   }
 }
