--- src/lib/schematics/form/files/__path__/__name@dasherize@if-flat__/__name@dasherize__.component.__styleext__
@@ -1,13 +1,0 @@
-.full-width {
-  width: 100%;
-}
-
-.shipping-card {
-  width: 500px;
-  margin: 20px auto;
-}
-
-mat-radio-button {
-  display: block;
-  margin: 5px 0;
-}
